Fix removal of unsaved calculation line in deleted state
Calling model.rollbackAttributes() on a new ember-data record
transitions the record to the deleted state.

+++ app/components/offerline/detail.js
@@ -57,7 +57,9 @@ export default class OfferlineDetailComponent extends Component {
   @task
   *deleteCalculationLine(calculationLine) {
     this.calculationLines.removeObject(calculationLine);
-    yield calculationLine.destroyRecord();
+    if (!calculationLine.isDeleted) {
+      yield calculationLine.destroyRecord();
+    }
   }
 
   @action
